requirements.txt
absl-py==2.0.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
altair==4.2.2
aniso8601==9.0.1
asgiref==3.6.0
astroid==2.14.2
astunparse==1.6.3
async-timeout==4.0.2
av==10.0.0
backports.zoneinfo==0.2.1
black==23.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
denoising-diffusion-pytorch==1.5.5
dill==0.3.6
dlib==19.24.1
dm-tree==0.1.8
einops==0.6.1
ema-pytorch==0.2.3
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.1.3
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-Migrate==2.5.2
flask-restx==0.5.1
Flask-SQLAlchemy==2.4.1
flatbuffers==23.5.26
fonttools==4.39.0
frozenlist==1.3.3
gast==0.3.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.59.0
h5py==2.10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.12.0
imutils==0.5.2
isort==5.12.0
itsdangerous==2.1.2
jax==0.4.13
jetson-emulator==1.0.0
Jinja2==3.1.2
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
libclang==14.0.6
lit==16.0.1
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.7.2
mccabe==0.7.0
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.22.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
onnx==1.12.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
protobuf==4.24.4
psutil==5.9.5
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.0
Pygments==2.14.0
pylint==2.16.2
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytorch-fid==0.3.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.4.1
seaborn==0.12.2
semver==3.0.0
six==1.15.0
smmap==5.0.0
SQLAlchemy==2.0.5.post1
stytch==6.7.0
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.0
tensorflow-addons==0.21.0
tensorflow-cpu==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.28.0
tensorflow-probability==0.21.0
tensorrt==0.0.1
termcolor==1.1.0
tflite-runtime==2.13.0
thop==0.1.1.post2209072238
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
torch==1.9.1
torch-tb-profiler==0.4.1
torchaudio==0.9.0a0+a85b239
torchvision==0.10.1
tornado==6.2
triton==2.0.0
typeguard==2.13.3
typing-extensions==3.7.4.3
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.4
validators==0.20.0
watchdog==3.0.0
Werkzeug==3.0.0
wrapt==1.12.1
yarl==1.8.2
zipp==3.15.0
git+https://github.com/ChaoningZhang/MobileSAM.git
boxmot